requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
sqlalchemy
langchain
langchain-community
langchain-google-genai
torch
transformers
beautifulsoup4
pypdf
pandas
llama-cpp-python
accelerate
hf_xet